refactor(core,query): share success responseType detection
